Re-refactored menu
