+++ app/src/main/java/com/zionhuang/music/ui/menu/PlayerSongMenu.kt
@@ -1,0 +1,303 @@
+package com.zionhuang.music.ui.menu
+
+import android.content.Intent
+import android.media.audiofx.AudioEffect
+import androidx.activity.compose.rememberLauncherForActivityResult
+import androidx.activity.result.contract.ActivityResultContracts
+import androidx.annotation.DrawableRes
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.WindowInsets
+import androidx.compose.foundation.layout.asPaddingValues
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.systemBars
+import androidx.compose.foundation.layout.width
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.window.DialogProperties
+import androidx.media3.common.PlaybackParameters
+import androidx.navigation.NavController
+import com.zionhuang.music.LocalPlayerConnection
+import com.zionhuang.music.R
+import com.zionhuang.music.db.entities.Event
+import com.zionhuang.music.models.MediaMetadata
+import com.zionhuang.music.ui.component.BigSeekBar
+import com.zionhuang.music.ui.component.BottomSheetState
+import com.zionhuang.music.ui.component.GridMenu
+import com.zionhuang.music.ui.component.GridMenuItem
+import kotlin.math.log2
+import kotlin.math.pow
+import kotlin.math.round
+
+@Composable
+fun PlayerSongMenu(
+    mediaMetadata: MediaMetadata,
+    navController: NavController,
+    playerBottomSheetState: BottomSheetState? = null,
+    event: Event? = null,
+    onShowDetailsDialog: () -> Unit = {},
+    onDismiss: () -> Unit
+) {
+    GenericSongMenu(
+        mediaMetadata = mediaMetadata,
+        navController = navController,
+        playerBottomSheetState = playerBottomSheetState,
+        event = event,
+        onDismiss = onDismiss,
+        topContent = { showExpanded: Boolean,
+                       onToggle: (Boolean) -> Unit
+            ->
+            PlayerSongMenuTopContent(
+                showPlayerItems = showExpanded,
+                toggle = onToggle
+            )
+        },
+        topContentPane = {
+            PlayerSongMenuTopContentPane(
+                onShowDetailsDialog = onShowDetailsDialog,
+                onDismiss = onDismiss
+            )
+        }
+    )
+}
+
+@Composable
+fun PlayerSongMenuTopContent(
+    showPlayerItems: Boolean,
+    toggle: (Boolean) -> Unit
+) {
+    val playerConnection = LocalPlayerConnection.current ?: return
+    val playerVolume = playerConnection.service.playerVolume.collectAsState()
+
+    Row(
+        horizontalArrangement = Arrangement.spacedBy(24.dp),
+        verticalAlignment = Alignment.CenterVertically,
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(horizontal = 24.dp)
+            .padding(top = 16.dp, bottom = 8.dp)
+    ) {
+        Icon(
+            painter = painterResource(R.drawable.volume_up),
+            contentDescription = null,
+            modifier = Modifier.size(28.dp)
+        )
+
+        BigSeekBar(
+            progressProvider = playerVolume::value,
+            onProgressChange = { playerConnection.service.playerVolume.value = it },
+            modifier = Modifier.weight(1f)
+        )
+
+        IconButton(
+            onClick = {
+                toggle(!showPlayerItems)
+            }
+        ) {
+            Icon(
+                painter = painterResource(if (showPlayerItems) R.drawable.expand_less else R.drawable.expand_more),
+                contentDescription = null,
+                modifier = Modifier.size(28.dp)
+            )
+        }
+    }
+}
+
+@Composable
+fun PlayerSongMenuTopContentPane(
+    onShowDetailsDialog: () -> Unit = {},
+    onDismiss: () -> Unit
+) {
+    val context = LocalContext.current
+    val playerConnection = LocalPlayerConnection.current ?: return
+    val activityResultLauncher =
+        rememberLauncherForActivityResult(ActivityResultContracts.StartActivityForResult()) { }
+
+    var showPitchTempoDialog by rememberSaveable {
+        mutableStateOf(false)
+    }
+
+    if (showPitchTempoDialog) {
+        PitchTempoDialog(
+            onDismiss = { showPitchTempoDialog = false }
+        )
+    }
+
+    GridMenu(
+        contentPadding = PaddingValues(
+            start = 8.dp,
+            end = 8.dp,
+        )
+    ) {
+        GridMenuItem(
+            icon = R.drawable.equalizer,
+            title = R.string.equalizer
+        ) {
+            val intent =
+                Intent(AudioEffect.ACTION_DISPLAY_AUDIO_EFFECT_CONTROL_PANEL).apply {
+                    putExtra(
+                        AudioEffect.EXTRA_AUDIO_SESSION,
+                        playerConnection.player.audioSessionId
+                    )
+                    putExtra(AudioEffect.EXTRA_PACKAGE_NAME, context.packageName)
+                    putExtra(AudioEffect.EXTRA_CONTENT_TYPE, AudioEffect.CONTENT_TYPE_MUSIC)
+                }
+            if (intent.resolveActivity(context.packageManager) != null) {
+                activityResultLauncher.launch(intent)
+            }
+            onDismiss()
+        }
+        GridMenuItem(
+            icon = R.drawable.tune,
+            title = R.string.advanced
+        ) {
+            showPitchTempoDialog = true
+        }
+        GridMenuItem(
+            icon = R.drawable.info,
+            title = R.string.details
+        ) {
+            onShowDetailsDialog()
+        }
+    }
+}
+
+@Composable
+fun PitchTempoDialog(
+    onDismiss: () -> Unit,
+) {
+    val playerConnection = LocalPlayerConnection.current ?: return
+    var tempo by remember {
+        mutableStateOf(playerConnection.player.playbackParameters.speed)
+    }
+    var transposeValue by remember {
+        mutableStateOf(round(12 * log2(playerConnection.player.playbackParameters.pitch)).toInt())
+    }
+    val updatePlaybackParameters = {
+        playerConnection.player.playbackParameters =
+            PlaybackParameters(tempo, 2f.pow(transposeValue.toFloat() / 12))
+    }
+
+    AlertDialog(
+        properties = DialogProperties(usePlatformDefaultWidth = false),
+        onDismissRequest = onDismiss,
+        dismissButton = {
+            TextButton(
+                onClick = {
+                    tempo = 1f
+                    transposeValue = 0
+                    updatePlaybackParameters()
+                }
+            ) {
+                Text(stringResource(R.string.reset))
+            }
+        },
+        confirmButton = {
+            TextButton(
+                onClick = onDismiss
+            ) {
+                Text(stringResource(android.R.string.ok))
+            }
+        },
+        text = {
+            Column {
+                ValueAdjuster(
+                    icon = R.drawable.slow_motion_video,
+                    currentValue = tempo,
+                    values = listOf(0.25f, 0.5f, 0.75f, 1f, 1.25f, 1.5f, 1.75f, 2f),
+                    onValueUpdate = {
+                        tempo = it
+                        updatePlaybackParameters()
+                    },
+                    valueText = { "x$it" },
+                    modifier = Modifier.padding(bottom = 12.dp)
+                )
+                ValueAdjuster(
+                    icon = R.drawable.discover_tune,
+                    currentValue = transposeValue,
+                    values = (-12..12).toList(),
+                    onValueUpdate = {
+                        transposeValue = it
+                        updatePlaybackParameters()
+                    },
+                    valueText = { "${if (it > 0) "+" else ""}$it" }
+                )
+            }
+        }
+    )
+}
+
+@Composable
+fun <T> ValueAdjuster(
+    @DrawableRes icon: Int,
+    currentValue: T,
+    values: List<T>,
+    onValueUpdate: (T) -> Unit,
+    valueText: (T) -> String,
+    modifier: Modifier = Modifier,
+) {
+    Row(
+        horizontalArrangement = Arrangement.spacedBy(24.dp),
+        verticalAlignment = Alignment.CenterVertically,
+        modifier = modifier
+    ) {
+        Icon(
+            painter = painterResource(icon),
+            contentDescription = null,
+            modifier = Modifier.size(28.dp)
+        )
+
+        IconButton(
+            enabled = currentValue != values.first(),
+            onClick = {
+                onValueUpdate(values[values.indexOf(currentValue) - 1])
+            }
+        ) {
+            Icon(
+                painter = painterResource(R.drawable.remove),
+                contentDescription = null
+            )
+        }
+
+        Text(
+            text = valueText(currentValue),
+            style = MaterialTheme.typography.titleMedium,
+            textAlign = TextAlign.Center,
+            modifier = Modifier.width(80.dp)
+        )
+
+        IconButton(
+            enabled = currentValue != values.last(),
+            onClick = {
+                onValueUpdate(values[values.indexOf(currentValue) + 1])
+            }
+        ) {
+            Icon(
+                painter = painterResource(R.drawable.add),
+                contentDescription = null
+            )
+        }
+    }
+}
