+++ app/src/main/java/com/zionhuang/music/ui/player/Queue.kt
@@ -82,7 +82,7 @@ import com.zionhuang.music.ui.component.BottomSheet
 import com.zionhuang.music.ui.component.BottomSheetState
 import com.zionhuang.music.ui.component.LocalMenuState
 import com.zionhuang.music.ui.component.MediaMetadataListItem
-import com.zionhuang.music.ui.menu.GenericSongMenu
+import com.zionhuang.music.ui.menu.PlayerSongMenu
 import com.zionhuang.music.utils.makeTimeString
 import com.zionhuang.music.utils.rememberPreference
 import kotlinx.coroutines.Dispatchers
@@ -320,12 +320,11 @@ fun Queue(
                 IconButton(
                     onClick = {
                         menuState.show {
-                            GenericSongMenu(
-                                mediaMetadata = mediaMetadata,
+                            PlayerSongMenu(
+                                mediaMetadata = mediaMetadata!!,
                                 navController = navController,
                                 playerBottomSheetState = playerBottomSheetState,
                                 onShowDetailsDialog = { showDetailsDialog = true },
-                                inPlayer = true,
                                 onDismiss = menuState::dismiss
                             )
                         }
