+++ app/src/main/java/com/zionhuang/music/ui/screens/HomeScreen.kt
@@ -38,7 +38,7 @@ import com.zionhuang.music.ui.component.NavigationTile
 import com.zionhuang.music.ui.component.NavigationTitle
 import com.zionhuang.music.ui.component.SongListItem
 import com.zionhuang.music.ui.component.YouTubeGridItem
-import com.zionhuang.music.ui.menu.GenericSongMenu
+import com.zionhuang.music.ui.menu.LibrarySongMenu
 import com.zionhuang.music.ui.menu.YouTubeAlbumMenu
 import com.zionhuang.music.ui.utils.SnapLayoutInfoProvider
 import com.zionhuang.music.utils.rememberPreference
@@ -172,7 +172,7 @@ fun HomeScreen(
                                         IconButton(
                                             onClick = {
                                                 menuState.show {
-                                                    GenericSongMenu(
+                                                    LibrarySongMenu(
                                                         mediaMetadata = song!!.toMediaMetadata(),
                                                         navController = navController,
                                                         onDismiss = menuState::dismiss
