+++ app/src/main/java/com/zionhuang/music/ui/screens/YouTubeBrowseScreen.kt
@@ -38,7 +38,7 @@ import com.zionhuang.music.ui.component.shimmer.ShimmerHost
 import com.zionhuang.music.ui.menu.YouTubeAlbumMenu
 import com.zionhuang.music.ui.menu.YouTubeArtistMenu
 import com.zionhuang.music.ui.menu.YouTubePlaylistMenu
-import com.zionhuang.music.ui.menu.GenericSongMenu
+import com.zionhuang.music.ui.menu.LibrarySongMenu
 import com.zionhuang.music.viewmodels.YouTubeBrowseViewModel
 
 @OptIn(ExperimentalMaterial3Api::class, ExperimentalFoundationApi::class)
@@ -91,7 +91,7 @@ fun YouTubeBrowseScreen(
                             onClick = {
                                 menuState.show {
                                     when (item) {
-                                        is SongItem -> GenericSongMenu(
+                                        is SongItem -> LibrarySongMenu(
                                             mediaMetadata = item.toMediaMetadata(),
                                             navController = navController,
                                             onDismiss = menuState::dismiss
