+++ app/src/main/java/com/zionhuang/music/ui/screens/artist/ArtistItemsScreen.kt
@@ -48,7 +48,7 @@ import com.zionhuang.music.ui.component.shimmer.ShimmerHost
 import com.zionhuang.music.ui.menu.YouTubeAlbumMenu
 import com.zionhuang.music.ui.menu.YouTubeArtistMenu
 import com.zionhuang.music.ui.menu.YouTubePlaylistMenu
-import com.zionhuang.music.ui.menu.GenericSongMenu
+import com.zionhuang.music.ui.menu.LibrarySongMenu
 import com.zionhuang.music.viewmodels.ArtistItemsViewModel
 
 @OptIn(ExperimentalFoundationApi::class, ExperimentalMaterial3Api::class)
@@ -110,7 +110,7 @@ fun ArtistItemsScreen(
                             onClick = {
                                 menuState.show {
                                     when (item) {
-                                        is SongItem -> GenericSongMenu(
+                                        is SongItem -> LibrarySongMenu(
                                             mediaMetadata = item.toMediaMetadata(),
                                             navController = navController,
                                             onDismiss = menuState::dismiss
@@ -203,7 +203,7 @@ fun ArtistItemsScreen(
                             onLongClick = {
                                 menuState.show {
                                     when (item) {
-                                        is SongItem -> GenericSongMenu(
+                                        is SongItem -> LibrarySongMenu(
                                             mediaMetadata = item.toMediaMetadata(),
                                             navController = navController,
                                             onDismiss = menuState::dismiss
