+++ app/src/main/java/com/zionhuang/music/ui/screens/artist/ArtistScreen.kt
@@ -82,11 +82,10 @@ import com.zionhuang.music.ui.component.shimmer.ButtonPlaceholder
 import com.zionhuang.music.ui.component.shimmer.ListItemPlaceHolder
 import com.zionhuang.music.ui.component.shimmer.ShimmerHost
 import com.zionhuang.music.ui.component.shimmer.TextPlaceholder
-import com.zionhuang.music.ui.menu.GenericSongMenu
+import com.zionhuang.music.ui.menu.LibrarySongMenu
 import com.zionhuang.music.ui.menu.YouTubeAlbumMenu
 import com.zionhuang.music.ui.menu.YouTubeArtistMenu
 import com.zionhuang.music.ui.menu.YouTubePlaylistMenu
-import com.zionhuang.music.ui.menu.GenericSongMenu
 import com.zionhuang.music.ui.utils.fadingEdge
 import com.zionhuang.music.ui.utils.resize
 import com.zionhuang.music.viewmodels.ArtistViewModel
@@ -227,7 +226,7 @@ fun ArtistScreen(
                                 IconButton(
                                     onClick = {
                                         menuState.show {
-                                            GenericSongMenu(
+                                            LibrarySongMenu(
                                                 mediaMetadata = song.toMediaMetadata(),
                                                 navController = navController,
                                                 onDismiss = menuState::dismiss
@@ -280,7 +279,7 @@ fun ArtistScreen(
                                     IconButton(
                                         onClick = {
                                             menuState.show {
-                                                GenericSongMenu(
+                                                LibrarySongMenu(
                                                     mediaMetadata = song.toMediaMetadata(),
                                                     navController = navController,
                                                     onDismiss = menuState::dismiss
@@ -334,7 +333,7 @@ fun ArtistScreen(
                                                 onLongClick = {
                                                     menuState.show {
                                                         when (item) {
-                                                            is SongItem -> GenericSongMenu(
+                                                            is SongItem -> LibrarySongMenu(
                                                                 mediaMetadata = item.toMediaMetadata(),
                                                                 navController = navController,
                                                                 onDismiss = menuState::dismiss
